PAX: support sum/count pb info combine
After PAX support the sum/count stats.
Then the pb stats combine function(which provider to storage_am) should also been updated.
Current change split the MicroPartitionStatisticsInfoCombine function to these parts:

  - The PrepareStatisticsInfoCombine function used to check the pb stats struct is valid.
  - The CommStatisticsInfoCombine function used to combine the required filed:
    count/hasnull/allnull.
  - The MinMaxStatisticsInfoCombine function used to combine the min/max.
  - The SumStatisticsInfoCombine function used to combine the sum.

+++ contrib/pax_storage/src/cpp/storage/orc/orc_reader.cc
@@ -34,7 +34,7 @@ class OrcGroupStatsProvider final : public ColumnStatsProvider {
         .colstats(column_index)
         .hasnull();
   }
-  uint32 NonNullRows(int column_index) const override {
+  uint64 NonNullRows(int column_index) const override {
     return format_reader_.file_footer_.stripes(group_index_)
         .colstats(column_index)
         .nonnullrows();
